scipy_2022/requirements.txt
flytekit>=1.1.0
flytekitplugins-deck-standard
flytekitplugins-pandera
flytekitplugins-spark
dataclasses_json
joblib
palmerpenguins
pandas
pyspark
scikit-learn
torch
